android-12.0.0_r2 part 1

+++ external/autotest/client/site_tests/hardware_UnsafeMemory/src/rowhammer-test-4d619293e1c7/rowhammer_test.cc
@@ -107,7 +107,7 @@ static void toggle(int iterations, int addr_count) {
       break;
     }
   }
-  t.print_iters(iterations * addr_count * toggles);
+  t.print_iters((uint64_t) iterations * addr_count * toggles);
 }
 
 void main_prog() {
